fix: limit number of replacements when injecting component
The issue presented with #21 was basically that we wanted to inject exactly once, but were calling `preg_replace` to perform the injection, which replaces for every match found by default.
An optional fourth argument to the function, however, allows you to specify an upper limit of how many replacements may be made.
I have set this to `1` throughout `AbstractInjector`, which has caused the previously failing test case to now pass, resolving the issue.

Fixes #21

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>