qa: Psalm issues
The majority of these are due to third-party packages not providing specific enough of return type hints, leading to "mixed assignment" and similar.
Others are due to usage of things like filter/reduce/walk whereby we are operating on a subset of all arguments, but the last one is required for the operation (unused closure argument).

I did refactor the PackageProviderDetectionFactory to be more clear about what it provides, and to ensure it never passes a null argument.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>