Prevent composer cache folder interaction in tests
Composer tries to create composer cache folder if it does not exist and
create protective .htaccess file in it.

Disable that with env variables for tests.

Composer silences warnings for when cache folder is not writable with
reduced error_reporting level but PHPUnit error handler is not affected
by the setting.

Signed-off-by: Aleksei Khudiakov <aleksey@xerkus.pro>